pyproject.toml
[project]
name = "kairos"
version = "0.1.0"
description = "Local AI-powered video clipping and story assembly platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "huey[sqlite]>=2.5.0",
    "yt-dlp>=2025.1.0",
    "apscheduler>=3.10.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["kairos*"]
